+++ authentication/src/main/java/project/goorm/authentication/business/core/domain/member/infrastructure/command/MemberJpaRepository.java
@@ -1,4 +1,4 @@
-package project.goorm.authentication.business.core.domain.member.infrastructure;
+package project.goorm.authentication.business.core.domain.member.infrastructure.command;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.goorm.authentication.business.core.domain.member.entity.Member;
